requirements.txt
google-generativeai
python-dotenv
pyyaml
python-hcl2
jinja2
